LayoutTests/http/tests/security/contentSecurityPolicy/register-bypassing-scheme-partial-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/resources/abe.png because it appears in neither the img-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: PASS (1/12)
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/resources/cssStyle.css because it appears in neither the style-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: PASS (2/12)
CONSOLE MESSAGE: PASS (3/12)
CONSOLE MESSAGE: PASS (4/12)
CONSOLE MESSAGE: FAIL (5/12)
CONSOLE MESSAGE: FAIL (6/12)
CONSOLE MESSAGE: PASS (7/12)
CONSOLE MESSAGE: FAIL (8/12)
CONSOLE MESSAGE: FAIL (9/12)
CONSOLE MESSAGE: PASS (10/12)
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/resources/abe.png because it appears in neither the img-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: PASS (11/12)
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/resources/cssStyle.css because it appears in neither the style-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: PASS (12/12)
This test ensures that registering a scheme as bypassing CSP actually bypasses CSP. This test passes if only PASSes are generated.
